removed loop when setting the value of a textbox in a label-spinbox
setting the value of a label-spinbox invoked the ref-set-hook
twice (via set-pvb-value which was also called when setting #_value of
the spinboxes text-box), leading to nasty loops. With the change,
changing a value of a label-spinbox (set-pvb-value) will now change
the #_value of the textbox which in turn will trigger
ref-set-pvb-value to actually trigger the propagation of the value
change of the label-spinbox to its model-ref and its ref-set-hook.